# openapi.yaml
openapi: 3.0.0
info:
  title: BankWiz API
  version: 0.1.0
  description: An API for accessing banking functionalities provided by BankWiz
servers:
  - url: 'http://localhost:8080/api'
...
paths:
...
  /group:
    post:
      tags:
        - Group
      summary: Create a group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreationRequest'
      responses:
        '200':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
  /group/groups:
    get:
      tags:
        - Group
      summary: Get my groups
      operationId: getGroups
      responses:
        '200':
          description: Successfully retrieved the list of groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
  '/group/{groupId}':
    get:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Get group information
      operationId: getGroup
      responses:
        '200':
          description: Successfully retrieved group information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
    put:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Update group information
      operationId: updateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateRequest'
      responses:
        '200':
          description: Group information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
    delete:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Delete a group
      operationId: deleteGroup
      responses:
        '200':
          description: Group deleted successfully
        '400':
          description: Invalid request. Please check the provided data.
  '/group/{groupId}/user':
    post:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Add a user to the group
      operationId: addUserToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUserGroupRequest'
      responses:
        '200':
          description: User added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
  '/group/{groupId}/user/{userId}':
    put:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Update user in a group
      operationId: updateUserInGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserGroupRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
    delete:
      tags:
        - Group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      summary: Remove user from a group
      operationId: removeUserFromGroup
      responses:
        '200':
          description: User removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
        '400':
          description: Invalid request. Please check the provided data.
...
components:
  schemas:
...
    GroupCreationRequest:
      type: object
      properties:
        groupName:
          type: string
      required:
        - groupName
    GroupUpdateRequest:
      type: object
      properties:
        groupName:
          type: string
    AddUserGroupRequest:
      type: object
      properties:
        userId:
          type: integer
        authorization:
          $ref: '#/components/schemas/GroupAuthorizationEnum'
      required:
        - userId
        - authorization
    UpdateUserGroupRequest:
      type: object
      properties:
        authorization:
          $ref: '#/components/schemas/GroupAuthorizationEnum'
...
    UserDTO:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: johndoe@example.com
      required:
        - userId
        - firstName
        - lastName
        - email
    UserGroupDTO:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserDTO'
        authorization:
          $ref: '#/components/schemas/GroupAuthorizationEnum'
      required:
        - user
        - authorization
    GroupDTO:
      type: object
      properties:
        groupId:
          type: integer
        groupName:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserGroupDTO'
      required:
        - groupId
        - groupName
        - users
...
    GroupAuthorizationEnum:
      type: string
      enum:
        - READ
        - WRITE
        - ADMIN